feat: Create timeseriesDetectorML class, perceptron subclass, and tests
This commit introduces a new abstract base class for time series machine learning detectors, `vlt.signal.timeseriesDetectorML.base`, and a concrete implementation using a perceptron, `vlt.signal.timeseriesDetectorML.perceptron`.

The base class defines a standard interface for training and evaluating time series detectors. It also includes a static helper method, `timeStamps2Observations`, for extracting training data from a time series, with options for peak finding.

The perceptron subclass provides a simple linear classifier for detecting patterns in time series data, including a bias term for improved accuracy.

Unit tests for both the base and perceptron classes are included in the `vlt.unittest.signal.timeSeriesDetectorML` package to ensure correctness.

This commit also includes fixes for bugs identified during testing, including a data type mismatch in the tests and a dimension mismatch in the perceptron evaluation method.

+++ +vlt/+unittest/+signal/+timeSeriesDetectorML/testPerceptron.m
@@ -9,7 +9,7 @@ classdef testPerceptron < matlab.unittest.TestCase
             learningRate = 0.05;
             p = vlt.signal.timeseriesDetectorML.perceptron(detectorSamples, learningRate);
 
-            testCase.verifyEqual(p.detectorSamples, detectorSamples);
+            testCase.verifyEqual(p.detectorSamples, uint64(detectorSamples));
             testCase.verifyEqual(p.learningRate, learningRate);
             testCase.verifyEqual(size(p.weights), [detectorSamples + 1, 1]);
         end
